refactor: alterar caminho de arquivos loas
