+++ src/modules/GetInformationFromSapienForSamir/loas/Business/Dossie/RestabelecimentoRequerimentosDossie.ts
@@ -1,32 +1,36 @@
-import { getXPathText } from "../../../../helps/GetTextoPorXPATH";
-import { arrayExisteCessadoOuSuspenso } from "./Help/ArrayExisteCessaoOuSuspenso";
-import { buscardatasLoas } from "./Help/BuscarDatas";
-import { EncontrarDataMaisAtual } from "./Help/EncontrarDataMaisAtual";
+import { getXPathText } from "../../../../../helps/GetTextoPorXPATH";
+import { arrayExisteCessadoOuSuspenso } from "../Help/ArrayExisteCessaoOuSuspenso";
+import { buscardatasLoas } from "../Help/BuscarDatas";
+import { EncontrarDataMaisAtual } from "../Help/EncontrarDataMaisAtual";
 
-export class RestabelecimentoRequerimentosSuperDossie{
+export class RestabelecimentoRequerimentos{
     async handle(parginaDosPrevFormatada: any):Promise<any>{
         //Estrutura para identificar a maior data, e fazer a subtração dela
-        let tamanhoColunasRequerimentos = 1;
+        let tamanhoColunasRequerimentos = 2;
         const arrayDatas: Array<Date> = [];
         let verificarWhileRequerimentos = true;
         while(verificarWhileRequerimentos){
-            if(typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[6]/table/tbody/tr[${tamanhoColunasRequerimentos}]`)) == 'object'){
+            if(typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[3]/table/tbody/tr[${tamanhoColunasRequerimentos}]`)) == 'object'){
                 verificarWhileRequerimentos = false; 
                 break;
             }
             tamanhoColunasRequerimentos++;
         }
-        const objetosEncontradosParaVerificar = []
-            for(let t=1; t<tamanhoColunasRequerimentos; t++){
-                if(typeof (getXPathText(parginaDosPrevFormatada,`/html/body/div/div[6]/table/tbody/tr[${t}]`)) === 'string'){
-                    const xpathColunaRequerimentos = `/html/body/div/div[6]/table/tbody/tr[${t}]`;
+        console.log("TAMANHO")
+        //console.log(tamanhoColunasRequerimentos)
+            const objetosEncontradosParaVerificar = []
+            for(let t=2; t<tamanhoColunasRequerimentos; t++){
+                if(typeof (getXPathText(parginaDosPrevFormatada,`/html/body/div/div[3]/table/tbody/tr[${t}]`)) === 'string'){
+                    const xpathColunaRequerimentos = `/html/body/div/div[3]/table/tbody/tr[${t}]`;
                     const xpathCoulaFormatadoRequerimentos: string = getXPathText(parginaDosPrevFormatada, xpathColunaRequerimentos);
+                    //console.log(buscardatasLoas(xpathCoulaFormatadoRequerimentos))
+                    //console.log(xpathCoulaFormatadoRequerimentos)
                     if(xpathCoulaFormatadoRequerimentos.indexOf("CESSADO") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("SUSPENSO") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("INDEFERIDO") !== -1){
                         if(xpathCoulaFormatadoRequerimentos.indexOf("87 - ") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("88 - ") !== -1){
-                            console.log("PASSSSOUUUU2")
+                            console.log("PASSSSOUUUU1")
                             if(xpathCoulaFormatadoRequerimentos.indexOf("CESSADO") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("SUSPENSO") !== -1){
                                 const buscarDataCessaoOuSuspenso = buscardatasLoas(xpathCoulaFormatadoRequerimentos);
-                                console.log(buscarDataCessaoOuSuspenso[0])
+                                //console.log(buscarDataCessaoOuSuspenso[0])
                                 if(!buscarDataCessaoOuSuspenso) return new Error("beneficio sem data")
                                 const restabelecimento = {
                                     beneficio: "cessaoOuSuspenso",
@@ -62,5 +66,4 @@ export class RestabelecimentoRequerimentosSuperDossie{
     
     }
 
-
