+++ src/modules/GetInformationFromSapienForSamir/loas/Business/SuperDossie/LoasEmpregoSuperDossie.ts
@@ -1,19 +1,21 @@
-import { getXPathText } from "../../../../helps/GetTextoPorXPATH"
-import { convertToDate } from "../../helps/createFormatDate";
-import { arrayExisteCessadoOuSuspenso } from "./Help/ArrayExisteCessaoOuSuspenso";
-import { buscarDataParaLoasEmprego } from "./Help/buscarDataParaLoasEmprego";
-import { buscardatasLoas } from "./Help/BuscarDatas";
-import { EncontrarDataMaisAtual } from "./Help/EncontrarDataMaisAtual";
 //iniciar a logica de buscar o emprego
 //primeiro buscar a data no ajuizamento
 //depois buscar a data nos requerimentos
 //depois buscar a data das previdenciarias
 //fazer comparacoes e tratar os erros possiveis
 
+import { getXPathText } from "../../../../../helps/GetTextoPorXPATH";
+import { convertToDate } from "../../../helps/createFormatDate";
+import { buscardatasLoas } from "../Help/BuscarDatas";
+import { EncontrarDataMaisAtual } from "../Help/EncontrarDataMaisAtual";
+import { buscarDataParaLoasEmprego } from "../Help/buscarDataParaLoasEmprego";
 
-export class LoasEmpregoDossie{
-    async execute(parginaDosPrevFormatada: any): Promise<boolean | object>{
 
+
+//organizar xpath para super dossies
+export class LoasEmpregoSuperDossie{
+    async execute(parginaDosPrevFormatada: any){
+        
         try{
 
             const xpathDataAjuzamento = "/html/body/div/div[4]/table/tbody/tr[2]/td"
@@ -23,22 +25,22 @@ export class LoasEmpregoDossie{
             if(dateAjuizamento.length == 0) new Error("data ajuizamento não encontrada");
             if(!(typeof(convertToDate(dateAjuizamento.trim())) == typeof(new Date()))) new Error("pegou xpath errado do ajuizamento");
 
-            // /html/body/div/div[6]/table/tbody/tr[1]
-            let tamanhoColunasRequerimentos = 2;
-        const arrayDatas: Array<Date> = [];
-        let verificarWhileRequerimentos = true;
-        while(verificarWhileRequerimentos){
-            if(typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[3]/table/tbody/tr[${tamanhoColunasRequerimentos}]`)) == 'object'){
-                verificarWhileRequerimentos = false; 
-                break;
+
+            let tamanhoColunasRequerimentos = 1;
+            const arrayDatas: Array<Date> = [];
+            let verificarWhileRequerimentos = true;
+            while(verificarWhileRequerimentos){
+                if(typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[6]/table/tbody/tr[${tamanhoColunasRequerimentos}]`)) == 'object'){
+                    verificarWhileRequerimentos = false; 
+                    break;
+                }
+                tamanhoColunasRequerimentos++;
             }
-            tamanhoColunasRequerimentos++;
-        }
 
             const objetosEncontradosParaVerificar = []
             for(let t=2; t<tamanhoColunasRequerimentos; t++){
-                if(typeof (getXPathText(parginaDosPrevFormatada,`/html/body/div/div[3]/table/tbody/tr[${t}]`)) === 'string'){
-                    const xpathColunaRequerimentos = `/html/body/div/div[3]/table/tbody/tr[${t}]`;
+                if(typeof (getXPathText(parginaDosPrevFormatada,`/html/body/div/div[6]/table/tbody/tr[${t}]`)) === 'string'){
+                    const xpathColunaRequerimentos = `/html/body/div/div[6]/table/tbody/tr[${t}]`;
                     const xpathCoulaFormatadoRequerimentos: string = getXPathText(parginaDosPrevFormatada, xpathColunaRequerimentos);
 
                     if(xpathCoulaFormatadoRequerimentos.indexOf("ATIVO") !== -1){
@@ -59,7 +61,6 @@ export class LoasEmpregoDossie{
                            const datasEncontradas = (buscardatasLoas(xpathCoulaFormatadoRequerimentos))
                         }
                       
-                    
                 }
             }
             
@@ -87,27 +88,27 @@ export class LoasEmpregoDossie{
             //buscar os dados agora ta tabela das "RELACOES PREVIDENCIARIAS"
 
 
-            let tamanhoColunaPrevidenciarias = 2;
+            let tamanhoColunaPrevidenciarias = 1;
             let verificarWhilePrevidenciarias = true;
-            while (verificarWhilePrevidenciarias) {
-                if (typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[4]/table/tbody/tr[${tamanhoColunaPrevidenciarias}]`)) == 'object') {
-                    verificarWhilePrevidenciarias = false;
+            while(verificarWhilePrevidenciarias){
+                if(typeof (getXPathText(parginaDosPrevFormatada, `html/body/div/div[7]/table/tbody/tr[${tamanhoColunaPrevidenciarias}]`)) == 'object'){
+                    verificarWhilePrevidenciarias = false; 
                     break;
                 }
                 tamanhoColunaPrevidenciarias++;
             }
             
-
-
+            // /html/body/div/div[7]/table/tbody/tr[1]/td[5]
+            // /html/body/div/div[7]/table/tbody/tr[2]/td[5]
             let dataInicioPrevidenciariasNaoExiste = false;
             let dataFimPrevidenciariasNaoExiste = false;
             let dataInicioEDataFimNaoExistem = false;
-            for(let i = 2; i<tamanhoColunaPrevidenciarias; i++){
-                const xpathDataInicioPrevidenciarias = `/html/body/div/div[4]/table/tbody/tr[${i}]/td[5]`
-                const xpathDataFimPrevidenciarias = `/html/body/div/div[4]/table/tbody/tr[${i}]/td[6]`
+            for(let i = 1; i<tamanhoColunaPrevidenciarias; i++){
+                const xpathDataInicioPrevidenciarias = `/html/body/div/div[7]/table/tbody/tr[${i}]/td[5]`
+                const xpathDataFimPrevidenciarias = `/html/body/div/div[7]/table/tbody/tr[${i}]/td[6]`
                 
-                let dataInicioPrevidenciaria = getXPathText(parginaDosPrevFormatada, xpathDataInicioPrevidenciarias);
-                let dataFimPrevidenciaria = getXPathText(parginaDosPrevFormatada, xpathDataFimPrevidenciarias);
+                const dataInicioPrevidenciaria = getXPathText(parginaDosPrevFormatada, xpathDataInicioPrevidenciarias);
+                const dataFimPrevidenciaria = getXPathText(parginaDosPrevFormatada, xpathDataFimPrevidenciarias);
 
                 
                 if(dataInicioPrevidenciaria.trim().length == 0 && dataFimPrevidenciaria.trim().length == 0){
@@ -141,6 +142,7 @@ export class LoasEmpregoDossie{
             }
 
             return false;
+
         }catch(e){
             console.log("ERRO EMPREGO LOAS" + e)
             return e
@@ -148,6 +150,7 @@ export class LoasEmpregoDossie{
 
 
 
+
+
     }
 }
-
