+++ src/modules/GetInformationFromSapienForSamir/loas/Business/SuperDossie/LoasLitispendenciaSuperDossie.ts
@@ -1,7 +1,7 @@
-import { getXPathText } from "../../../../helps/GetTextoPorXPATH";
-import { arrayExisteCessadoOuSuspenso } from "./Help/ArrayExisteCessaoOuSuspenso";
-import { buscardatasLoas } from "./Help/BuscarDatas";
-import { EncontrarDataMaisAtual } from "./Help/EncontrarDataMaisAtual";
+import { getXPathText } from "../../../../../helps/GetTextoPorXPATH";
+import { arrayExisteCessadoOuSuspenso } from "../Help/ArrayExisteCessaoOuSuspenso";
+import { buscardatasLoas } from "../Help/BuscarDatas";
+import { EncontrarDataMaisAtual } from "../Help/EncontrarDataMaisAtual";
 
 export class LoasLitispendenciaSuperDossie{
     async handle(parginaDosPrevFormatada: any):Promise<any>{
