+++ src/modules/GetInformationFromSapienForSamir/loas/Business/SuperDossie/RestabelecimentoRequerimentosSuperDossie.ts
@@ -1,36 +1,32 @@
-import { getXPathText } from "../../../../helps/GetTextoPorXPATH";
-import { arrayExisteCessadoOuSuspenso } from "./Help/ArrayExisteCessaoOuSuspenso";
-import { buscardatasLoas } from "./Help/BuscarDatas";
-import { EncontrarDataMaisAtual } from "./Help/EncontrarDataMaisAtual";
+import { getXPathText } from "../../../../../helps/GetTextoPorXPATH";
+import { arrayExisteCessadoOuSuspenso } from "../Help/ArrayExisteCessaoOuSuspenso";
+import { buscardatasLoas } from "../Help/BuscarDatas";
+import { EncontrarDataMaisAtual } from "../Help/EncontrarDataMaisAtual";
 
-export class RestabelecimentoRequerimentos{
+export class RestabelecimentoRequerimentosSuperDossie{
     async handle(parginaDosPrevFormatada: any):Promise<any>{
         //Estrutura para identificar a maior data, e fazer a subtração dela
-        let tamanhoColunasRequerimentos = 2;
+        let tamanhoColunasRequerimentos = 1;
         const arrayDatas: Array<Date> = [];
         let verificarWhileRequerimentos = true;
         while(verificarWhileRequerimentos){
-            if(typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[3]/table/tbody/tr[${tamanhoColunasRequerimentos}]`)) == 'object'){
+            if(typeof (getXPathText(parginaDosPrevFormatada, `/html/body/div/div[6]/table/tbody/tr[${tamanhoColunasRequerimentos}]`)) == 'object'){
                 verificarWhileRequerimentos = false; 
                 break;
             }
             tamanhoColunasRequerimentos++;
         }
-        console.log("TAMANHO")
-        console.log(tamanhoColunasRequerimentos)
-            const objetosEncontradosParaVerificar = []
-            for(let t=2; t<tamanhoColunasRequerimentos; t++){
-                if(typeof (getXPathText(parginaDosPrevFormatada,`/html/body/div/div[3]/table/tbody/tr[${t}]`)) === 'string'){
-                    const xpathColunaRequerimentos = `/html/body/div/div[3]/table/tbody/tr[${t}]`;
+        const objetosEncontradosParaVerificar = []
+            for(let t=1; t<tamanhoColunasRequerimentos; t++){
+                if(typeof (getXPathText(parginaDosPrevFormatada,`/html/body/div/div[6]/table/tbody/tr[${t}]`)) === 'string'){
+                    const xpathColunaRequerimentos = `/html/body/div/div[6]/table/tbody/tr[${t}]`;
                     const xpathCoulaFormatadoRequerimentos: string = getXPathText(parginaDosPrevFormatada, xpathColunaRequerimentos);
-                    //console.log(buscardatasLoas(xpathCoulaFormatadoRequerimentos))
-                    console.log(xpathCoulaFormatadoRequerimentos)
                     if(xpathCoulaFormatadoRequerimentos.indexOf("CESSADO") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("SUSPENSO") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("INDEFERIDO") !== -1){
                         if(xpathCoulaFormatadoRequerimentos.indexOf("87 - ") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("88 - ") !== -1){
-                            console.log("PASSSSOUUUU1")
+                            console.log("----LOOOOP")
                             if(xpathCoulaFormatadoRequerimentos.indexOf("CESSADO") !== -1 || xpathCoulaFormatadoRequerimentos.indexOf("SUSPENSO") !== -1){
                                 const buscarDataCessaoOuSuspenso = buscardatasLoas(xpathCoulaFormatadoRequerimentos);
-                                console.log(buscarDataCessaoOuSuspenso[0])
+                                console.log("---DATA: " + buscarDataCessaoOuSuspenso[0])
                                 if(!buscarDataCessaoOuSuspenso) return new Error("beneficio sem data")
                                 const restabelecimento = {
                                     beneficio: "cessaoOuSuspenso",
@@ -56,6 +52,9 @@ export class RestabelecimentoRequerimentos{
                     }
                 }
             }
+
+            console.log("---DATA MAIS ATUAL: " + EncontrarDataMaisAtual(objetosEncontradosParaVerificar).data)
+
             if(objetosEncontradosParaVerificar.length == 0) return false
             if(!arrayExisteCessadoOuSuspenso(objetosEncontradosParaVerificar)) return false
             if(EncontrarDataMaisAtual(objetosEncontradosParaVerificar).beneficio == "cessaoOuSuspenso") return true
@@ -66,4 +65,5 @@ export class RestabelecimentoRequerimentos{
     
     }
 
+
